requirements.txt
asgiref==3.2.10
Django==3.1.1
djangorestframework==3.11.1
icalendar==5.0.3
python-dateutil==2.8.2
pytz==2020.1
six==1.16.0
sqlparse==0.3.1
psycopg2-binary==2.8.6